Scala 2.13 upgrade (#70)
* cleanup types and collections

* correct tests

* remove warnings

* remove the unicode arrow

* more changes

* bug

* fix strings

* remove generated classes

* allow generated to be stored in a known directory

* bugs

* Remove unicode comments

* FIx rebase weirdness

* some clean up

* fix missing ssl code

* remove compat lib

+++ burst-fabric/src/main/scala/org/burstsys/fabric/data/model/slice/region/hose/hose.scala
@@ -69,7 +69,11 @@ package object hose extends VitalsLogger with FabricRegionInflator {
     /**
      * @return the ratio of inflated to deflated bytes processed
      */
-    final def compressionRatio: Double = if(deflatedByteCount==0) 0 else inflatedByteCount / deflatedByteCount
+    final def compressionRatio: Double =
+      if(deflatedByteCount==0)
+        0.0
+      else
+        inflatedByteCount.toDouble / deflatedByteCount
 
     def sprayMetrics: String =
       s"""|   parcelCount=$parcelCount (${prettySizeString(parcelCount)})
